packages/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.8.2
requests>=2.31.0
python-dateutil>=2.8.2